schitts/lecarrou/app/requirements/base.txt
amqp==5.0.2
beautifulsoup4==4.9.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==2.2.5
django-bootstrap4==1.0.1
django-celery-beat==2.1.0
django-cors-headers==3.5.0
django-crispy-forms==1.7.2
django-dbbackup==3.3.0
django-debug-toolbar==2.0
django-extensions==2.2.9
django-partial-date==1.2.2
django-safedelete==0.5.2
django-simple-history==2.7.3
django-widget-tweaks==1.4.5
djangorestframework==3.11.1
Faker==5.0.2
idna==2.10
kombu==5.0.2
Pillow==6.2.1
prompt-toolkit==3.0.8
psycopg2==2.8.3
python-dateutil==2.8.1
python-gettext==4.0
pytz==2019.2
randomuser==1.6
redis==3.5.3
reportlab==3.5.32
requests==2.25.1
selenium==3.141.0
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.6
vine==5.0.0
wcwidth==0.2.5
xlwt==1.3.0